# fix: Disables implicit Cookies url decoding