--- a/internal/corazawaf/transaction.go
+++ b/internal/corazawaf/transaction.go
@@ -321,7 +321,7 @@ func (tx *Transaction) AddRequestHeader(key string, value string) {
 		}
 	case "cookie":
 		// Cookies use the same syntax as GET params but with semicolon (;) separator
-		// noUrlUnescape is used to avoid implicitly performing an URL decode on the cookies
+		// WithoutUnescape is used to avoid implicitly performing an URL decode on the cookies
 		values := urlutil.ParseQueryWithoutUnescape(value, ';')
 		for k, vr := range values {
 			for _, v := range vr {