--- a/internal/url/url.go
+++ b/internal/url/url.go
@@ -52,27 +52,13 @@ func QueryUnescape(input string) string {
 				res.WriteByte(ci)
 				continue
 			}
-			hi := input[i+1]
-			lo := input[i+2]
-			switch {
-			case hi >= '0' && hi <= '9':
-				hi -= '0'
-			case hi >= 'a' && hi <= 'f':
-				hi -= 'a' - 10
-			case hi >= 'A' && hi <= 'F':
-				hi -= 'A' - 10
-			default:
+			hi, ok := hexDigitToByte(input[i+1])
+			if !ok {
 				res.WriteByte(ci)
 				continue
 			}
-			switch {
-			case lo >= '0' && lo <= '9':
-				lo -= '0'
-			case lo >= 'a' && lo <= 'f':
-				lo -= 'a' - 10
-			case lo >= 'A' && lo <= 'F':
-				lo -= 'A' - 10
-			default:
+			lo, ok := hexDigitToByte(input[i+2])
+			if !ok {
 				res.WriteByte(ci)
 				continue
 			}
@@ -83,4 +69,17 @@ func QueryUnescape(input string) string {
 		res.WriteByte(ci)
 	}
 	return res.String()
+}
+
+func hexDigitToByte(digit byte) (byte, bool) {
+	switch {
+	case digit >= '0' && digit <= '9':
+		return digit - '0', true
+	case digit >= 'a' && digit <= 'f':
+		return digit - 'a' + 10, true
+	case digit >= 'A' && digit <= 'F':
+		return digit - 'A' + 10, true
+	default:
+		return 0, false
+	}
 }