… go.sum
… go.work.sum
… testing/coreruleset/go.sum