Tentatively closes https://github.com/corazawaf/coraza/issues/920.