@jcchavezs 2023-11-27
> I still think we split the function instead of using using a Boolean constant

Kind of agree but that would require to copy the code. Another option could be:

```
type kvEscape func(string) string

var (
	urlUnescape  kvEscape = QueryUnescape
	NoEscape      kvEscape = func(s string) { return s }
)

...

q := ParseQuery(out, '&', urlUnescape)
```