@anuraaga 2023-11-28
We could split the API with private function

```
func ParseQuery(string)
func ParseQueryWithoutUnescape(string)

func doParseQuery(string, bool)
```