@M4tteoP 2023-11-28
Thank you all for the inputs!
I went for the `doParseQuery` avoiding public booleans and I did a minor refactor to `QueryUnescape` to improve a bit the readability. I can not notice meaningful differences in benchmarks:

**QueryUnescape**
Before:
```
goos: darwin
goarch: arm64
pkg: github.com/corazawaf/coraza/v3/internal/url
BenchmarkQueryUnescape-10    	 4639016	       252.8 ns/op	      78 B/op	       7 allocs/op
PASS
ok  	github.com/corazawaf/coraza/v3/internal/url	1.526s
```

After:
```
goos: darwin
goarch: arm64
pkg: github.com/corazawaf/coraza/v3/internal/url
BenchmarkQueryUnescape-10    	 4720306	       249.2 ns/op	      78 B/op	       7 allocs/op
PASS
ok  	github.com/corazawaf/coraza/v3/internal/url	1.515s
```

**ParseQuery** (same QueryUnescape)
Before:
```
goos: darwin
goarch: arm64
pkg: github.com/corazawaf/coraza/v3/internal/url
BenchmarkParseQuery-10    	 2169714	       544.6 ns/op	     627 B/op	       5 allocs/op
PASS
ok  	github.com/corazawaf/coraza/v3/internal/url	1.835s
```
After:
```
goos: darwin
goarch: arm64
pkg: github.com/corazawaf/coraza/v3/internal/url
BenchmarkParseQuery-10    	 2197615	       544.0 ns/op	     627 B/op	       5 allocs/op
PASS
ok  	github.com/corazawaf/coraza/v3/internal/url	1.842s
```