@M4tteoP 2023-12-11
Merging, mind that in the next release, I would write a line about this engine behavior change